backend_service/requirements.txt
fastapi
uvicorn
openai
pypdf
pandas
python-multipart
python-dotenv
pytest
httpx
langgraph 
langchain 
langchain-openai
matplotlib
networkx